refactor. 정보수정 모달 component 관심사 분리 #33
