+++ src/components/HeaderAside.tsx
@@ -6,9 +6,9 @@ import theme from '../shared/style/theme';
 import * as t from '../style/headerAside.style';
 import GlobalModal from './common/GlobalModal';
 import { SideMenuDrop } from './common/MenuDrop';
+import MyUserInfoModal from './common/UserInfoModal';
 import type { CateType } from './Header';
 import { cateData } from './Header';
-import MyUserInfoModal from './MyUserInfoModal';
 
 export default function HeaderAside() {
   const { isDropped, dropRef, handleRemove } = useDropDown();
